+++ solstice/src/main/java/dev/equo/ide/IdeLockFile.java
@@ -20,7 +20,7 @@ import java.util.ArrayList;
 import javax.annotation.Nullable;
 
 public class IdeLockFile {
-	private File workspaceDir;
+	private final File workspaceDir;
 
 	private IdeLockFile(File workspaceDir) {
 		this.workspaceDir = workspaceDir;
@@ -72,7 +72,8 @@ public class IdeLockFile {
 
 	long readPidToken() {
 		return FileMisc.readToken(workspaceDir, PID_FILENAME)
-				.map(str -> str.isEmpty() ? NO_TOKEN_FILE : Long.parseLong(str))
+				.filter(str -> !str.isEmpty())
+				.map(Long::parseLong)
 				.orElse(NO_TOKEN_FILE);
 	}